+++ core/src/index.ts
@@ -4,3 +4,4 @@ export * from "./events/index.js";
 export * from "./okf/index.js";
 export * from "./emit/index.js";
 export * from "./adapters/index.js";
+export * from "./fs/index.js";